pyproject.toml
[tool.poetry]
name = "gpt4-slack-bot"
version = "0.1.0"
description = ""
authors = ["Dmytro Sorochynskyi <dmytro.sorochynskyi@opower.com>"]
readme = "README.md"
packages = [{include = "gpt4_slack_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
openai = "^0.27.2"
slackclient = "^2.9.4"
Flask = "^2.2.3"
slackeventsapi = "3.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
